move existing tests of redis/results backend after refactor
